feat: moved chat into its own reusable component
