+++ src/app/owner/messaging/chat/page.module.scss
@@ -6,75 +6,6 @@
     height: 84vh;
     gap: 16px;
 
-    .chat {
-        height: 100%;
-        display: flex;
-        flex-direction: column;
-    }
-
-    .chat_header {
-        background-color: $color-off-white;
-        padding: 20px;
-        border-radius: 20px;
-    }
-
-    .chat_header_name {
-        color: $color-black;
-        font-size: $font-med-large;
-        font-weight: bold;
-        margin: 0;
-    }
-
-    .chat_subheader {
-        color: $color-gray;
-        font-size: $font-small-med;
-    }
-
-    .chat_body {
-        overflow-y: auto;
-        display: flex;
-        flex: 1;
-        flex-direction: column;
-        gap: 0.5rem;
-        padding: 1rem;
-    }
-
-    .chat_input_container {
-        height: 15%;
-        background: $color-blue;
-        display: flex;
-        justify-content: flex-end;
-    }
-
-    .input {
-        border: 2px solid $color-green-dark;
-        border-radius: 20px;
-    }
-
-    // ----- Other person's messages -----
-    .message_other {
-        align-self: flex-start;
-        max-width: 70%;
-        background: $color-off-white;
-        color: $color-gray;
-        padding: 0.75rem 1rem;
-        border-radius: 12px;
-        border-top-left-radius: 4px; // subtle "speech bubble" shape
-        font-size: 0.9rem;
-    }
-
-    // ----- Your messages -----
-    .message_me {
-        align-self: flex-end;
-        max-width: 70%;
-        background: $color-green; // soft green similar to your UI
-        color: $color-gray;
-        padding: 0.75rem 1rem;
-        border-radius: 12px;
-        border-top-right-radius: 4px;
-        font-size: 0.9rem;
-    }
-
     .pet_profile {
         width: 350px;
         overflow-y: auto;
@@ -83,7 +14,7 @@
     }
 }
 
-@media (max-width: 800px) {
+@media (max-width: 900px) {
     .page {
         grid-template-columns: 1fr; // only show the chat
     }
