+++ src/models/vet.ts
@@ -5,6 +5,7 @@ export class Vet {
     public readonly password: string;
     public readonly id: string;
     public readonly certification: string;
+    public readonly token: string;
 
     constructor(JSON) {
         Object.assign(this, JSON);
